Allow unconverted Params
Love this library, and I'd like to be able to use it with a [django-rest-framework-filters backend](https://github.com/philipn/django-rest-framework-filters).  However, the double__underscore thing it does for filtering breaks on the params filter.

It'd be great to have a way to pass arguments I don't want to send through a case converter.